Moved all plugins keybinds to their respective config files

+++ nvim-plugins-config/registers.nvim.lua
@@ -1,6 +1,7 @@
 require'registers'.setup {
 	window = {
-		border = "single",
+		border = "double",
 		transparency = 0,
 	},
 }
+
